+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Development
 - Add support for wetterdienst core API in cli and restapi
 - Export: Use InfluxDBClient instead of DataFrameClient and improve connection handling with InfluxDB 1.x
 - Export: Add support for InfluxDB 2.x
+- Fix InfluxDB export by skipping empty fields
 
 0.19.0 (14.05.2021)
 *******************